Fix 위도와 경도 타입 double로 변경

+++ semi-project/src/main/java/com/spring/semi/main/Entity/Property.java
@@ -17,6 +17,6 @@ public class Property {
 	private int realestateNo;
 	private String adstrdNm;
 	private int excheGtn;
-	private int latitude;
-	private int longitude;
+	private double latitude;
+	private double longitude;
 }
